pyproject.toml
[project]
name = "CineBaianos-bot"
version = "1.3.0"
description = "Bot do Discord para registrar filmes e votos dos Cinebaianos, com integração com Google Sheets."
authors = [ { name = "Bruno Reolon", email = "bruno.reolonn@email.com" } ]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py==2.5.2",
    "gspread==6.2.1",
    "oauth2client==4.1.3",
    "requests==2.32.4",
    "python-dotenv==1.1.1",
    "tmdbv3api==1.9.0",
    "google-auth==2.40.3",
    "gspread-formatting==1.2.1"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]